requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
faker==20.1.0
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0